+++ acme/resources/ACP.py
@@ -88,7 +88,10 @@ class ACP(AnnounceableResource):
 			acpi = r.acpi
 			if self.ri in acpi:
 				acpi.remove(self.ri)
-				r['acpi'] = acpi
+				if len(acpi) > 0:
+					r['acpi'] = acpi
+				else:
+					r['acpi'] = None	# Remove acpi from resource if empty
 				r.dbUpdate()